requirements.txt
bcrypt==4.2.0
PyMySQL==1.1.1
Jinja2==3.1.4
python-dotenv==1.0.1
